+++ jetty-http3/http3-common/src/main/java/org/eclipse/jetty/http3/internal/parser/ControlParser.java
@@ -136,7 +136,7 @@ public class ControlParser
         }
     }
 
-    private void sessionFailure(ByteBuffer buffer, int error, String reason)
+    private void sessionFailure(ByteBuffer buffer, long error, String reason)
     {
         unknownBodyParser.sessionFailure(buffer, error, reason);
     }
